bot/handlers.py
from .actions import *

from telegram.ext import filters, CommandHandler, MessageHandler

start_handler = CommandHandler('start', start)
echo_handler = MessageHandler(filters.TEXT & (~filters.COMMAND), echo)